fix an implicit conversion issue

+++ cudax/test/stf/loop_dispatch/nested_loop_dispatch.cu
@@ -38,7 +38,7 @@ int main()
       auto lA = ctx.logical_data<int>(size_t(1024 * 1024));
 
       ctx.parallel_for(ctx.current_exec_place(), lA.shape(), lA.write())->*[] __device__(size_t i, auto a) {
-        a(i) = (int) (10.0 * cuda::std::cos(i));
+        a(i) = (int) (10.0 * cuda::std::cos((double)i));
       };
 
       ctx.parallel_for(ctx.current_exec_place(), lA.shape(), lA.rw(), lB.read())
